Extract query option reader helpers
